Small refactoring II.
