+++ lib/content-services/content-metadata/components/content-metadata-card/content-metadata-card.component.spec.ts
@@ -25,6 +25,7 @@ import { ContentMetadataComponent } from '../content-metadata/content-metadata.c
 import { MatExpansionModule, MatCardModule, MatButtonModule, MatIconModule } from '@angular/material';
 import { ContentMetadataService } from '../../services/content-metadata.service';
 import { BasicPropertiesService } from '../../services/basic-properties.service';
+import { GroupedPropertiesService } from '../../services/grouped-properties.service';
 import { PropertyDescriptorLoaderService } from '../../services/properties-loader.service';
 import { PropertyDescriptorsService } from '../../services/property-descriptors.service';
 import { AspectWhiteListService } from '../../services/aspect-whitelist.service';
@@ -52,6 +53,7 @@ describe('ContentMetadataCardComponent', () => {
             providers: [
                 ContentMetadataService,
                 BasicPropertiesService,
+                GroupedPropertiesService,
                 PropertyDescriptorLoaderService,
                 PropertyDescriptorsService,
                 AspectWhiteListService,
