+++ lib/content-services/content-metadata/components/content-metadata/content-metadata.component.spec.ts
@@ -25,6 +25,7 @@ import { ContentMetadataComponent } from './content-metadata.component';
 import { MatExpansionModule, MatButtonModule, MatIconModule } from '@angular/material';
 import { ContentMetadataService } from '../../services/content-metadata.service';
 import { BasicPropertiesService } from '../../services/basic-properties.service';
+import { GroupedPropertiesService } from '../../services/grouped-properties.service';
 import { PropertyDescriptorLoaderService } from '../../services/properties-loader.service';
 import { PropertyDescriptorsService } from '../../services/property-descriptors.service';
 import { AspectWhiteListService } from '../../services/aspect-whitelist.service';
@@ -53,6 +54,7 @@ describe('ContentMetadataComponent', () => {
             providers: [
                 ContentMetadataService,
                 BasicPropertiesService,
+                GroupedPropertiesService,
                 PropertyDescriptorLoaderService,
                 PropertyDescriptorsService,
                 AspectWhiteListService,
